requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-JWT-Extended==4.6.0
FormEncode==2.1.0
fpdf==1.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
PyDispatcher==2.0.7
PyJWT==2.9.0
SQLObject==3.11.0
Werkzeug==3.0.4
